requirements.txt
playwright==1.58.0
beautifulsoup4
requests
azure-storage-blob